requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
python-dotenv==0.18.0
pydantic==1.8.2
ormar==0.10.13
asyncpg==0.23.0
psycopg2-binary==2.9.1
databases==0.4.3
SQLAlchemy==1.3.23
email-validator==1.1.3
python-jose==3.2.0
passlib[bcrypt]==1.7.4
Jinja2==3.0.1
google-api-python-client==2.11.0
pycoingecko==2.2.0
httpx==0.18.2
web3==5.20.1
ciso8601==2.1.3
python-multipart==0.0.5